Reorder scene assets per convention document & add reloader instead of reloading scene in UI
