+++ Assets/Prefabs/Managers/Encompasser.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 05c2ac9b2d6748742a13493b906b4c2a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
