+++ Assets/Prefabs/Managers/Loader.prefab
@@ -45,3 +45,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   input: {fileID: 11400000, guid: b165bc5393b36f14d8538a3f4673158a, type: 2}
+  onPlayerDeath: {fileID: 11400000, guid: 22935ebca27667842b100ca2831ca89e, type: 2}
+  onRestart: {fileID: 11400000, guid: c18d715c1cfa3df4cb5bdc4a92fb77ce, type: 2}
+  panelDuration: 1.5
+  onLoadScene: {fileID: 0}
