+++ Assets/Prefabs/UI/UI.prefab
@@ -1504,9 +1504,8 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   gameOverPanel: {fileID: 1234823667071641774}
-  panelDuration: 1.5
-  loader: {fileID: 0}
   onPlayerDeath: {fileID: 11400000, guid: 22935ebca27667842b100ca2831ca89e, type: 2}
+  onRestart: {fileID: 11400000, guid: c18d715c1cfa3df4cb5bdc4a92fb77ce, type: 2}
   onDoorPosition: {fileID: 11400000, guid: a30e8b5ac8160044b8a8c1e7b00ec593, type: 2}
   goObject: {fileID: 5710297127132402157}
   maxGoPositions: {x: 800, y: 400}
