+++ Assets/Prefabs/UI/UILvl1.prefab
@@ -24,6 +24,10 @@ PrefabInstance:
       propertyPath: m_AnchorMax.y
       value: 0
       objectReference: {fileID: 0}
+    - target: {fileID: 3388581458380239016, guid: bb1e495a18126e64fb441c85a9be2fae, type: 3}
+      propertyPath: onRestart
+      value: 
+      objectReference: {fileID: 11400000, guid: c18d715c1cfa3df4cb5bdc4a92fb77ce, type: 2}
     - target: {fileID: 3949237283212773107, guid: bb1e495a18126e64fb441c85a9be2fae, type: 3}
       propertyPath: m_AnchorMax.x
       value: 1
@@ -46,7 +50,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 6896562840811599013, guid: bb1e495a18126e64fb441c85a9be2fae, type: 3}
       propertyPath: m_Name
-      value: UI Variant
+      value: UILvl1
       objectReference: {fileID: 0}
     - target: {fileID: 8485514636649728380, guid: bb1e495a18126e64fb441c85a9be2fae, type: 3}
       propertyPath: m_Pivot.x
