+++ Assets/Scenes/Levels/Level1/NavMesh-NavMesh Surface 2.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 77a31213eb1e51842b5a02c7e1dc6277
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 23800000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
