+++ Assets/Scenes/Levels/Level2/NavMesh-NavMesh Surface 1.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b0d7b185f13c8304f858223a587ac5ea
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 23800000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
