+++ Assets/Scenes/Levels/Level2/NavMesh-NavMesh Surface.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 26334be8b089845498d15313b93b3ce4
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 23800000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
