+++ Assets/Scenes/Levels/Level3/NavMesh-NavMesh Surface.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4fd603b8a654d474d954c8cf7e322c4a
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 23800000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
