+++ Assets/Scenes/Managers/Managers.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 3af43bfe61bd8ae40b9c52e7d87de683
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
