+++ Assets/ScriptableObjects/Events/Scenes/OnRestart.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c18d715c1cfa3df4cb5bdc4a92fb77ce
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
