--- Assets/Scripts/Managers/Loader.cs
@@ -1,23 +1,0 @@
-using Events.Scriptables;
-using Player;
-using UnityEngine;
-using UnityEngine.SceneManagement;
-
-namespace Managers
-{
-    public class Loader : MonoBehaviour
-    {
-        [SerializeField] private InputHandler input;
-        [SerializeField] private StringEventChannelSO onLoadScene;
-        
-        void Start()
-        {
-            input.OnRestartScene.AddListener(RestartScene);
-        }
-
-        public void RestartScene()
-        {
-            onLoadScene?.RaiseEvent(SceneManager.GetActiveScene().name);
-        }
-    }
-}
